An ability to load files in task.

+++ Languages/LanguageExecutor.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using JobeSharp.Languages.Abstract;
 using JobeSharp.Sandbox;
+using JobeSharp.Services;
 
 namespace JobeSharp.Languages
 {
@@ -11,22 +12,33 @@ namespace JobeSharp.Languages
             Path.Combine(Path.GetTempPath(), "jobe", "executions", Guid.NewGuid().ToString());
         
         private ExecutionTask Task { get; }
+        private FileCache FileCache { get; }
 
-        public LanguageExecutor(ExecutionTask task)
+        public LanguageExecutor(ExecutionTask task, FileCache fileCache)
         {
             Task = task;
+            FileCache = fileCache;
 
             Directory.CreateDirectory(WorkTempDirectory);
         }
 
         public ExecutionResult Execute()
         {
+            LoadFiles();
             return Task.Language switch
             {
                 IInterpreted interpreted => ExecuteInterpreted(interpreted),
                 ICompiled compiled => ExecuteCompiled(compiled),
                 _ => throw new NotImplementedException()
             };
+        }
+
+        private void LoadFiles()
+        {
+            foreach (var (id, path) in Task.CachedFilesIdPath)
+            {
+                File.WriteAllBytes(Path.Combine(WorkTempDirectory, path), FileCache.ReadBytes(id));
+            }
         }
 
         private ExecutionResult ExecuteInterpreted(IInterpreted interpreted)
@@ -58,6 +70,11 @@ namespace JobeSharp.Languages
 
         public void Dispose()
         {
+            if (Task.Debug)
+            {
+                return;
+            }
+            
             Directory.Delete(WorkTempDirectory, recursive: true);
         }
     }
